Clenup in UndoStackItem
Made description fixed length buffer in Class
Moved some function definitions to header
Changed return type of get_size
Fixed size calculations.

+++ cinelerra/mainundo.h
@@ -87,7 +87,7 @@ public:
 
 	void set_data_before(const char *data);
 	virtual void undo();
-	virtual int get_size();
+	virtual size_t get_size();
 
 private:
 // type of modification
